requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
defusedxml==0.7.1
diff-match-patch==20230430
dj-rest-auth==2.2.8
Django==3.2.25
django-apscheduler==0.6.2
django-cors-headers==3.13.0
django-environ==0.11.2
django-filter==22.1
django-import-export==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
idna==3.8
odfpy==1.4.1
openpyxl==3.1.5
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.2
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tablib==3.6.1
tzlocal==5.2
urllib3==2.2.2
xlrd==2.0.1
xlwt==1.3.0
